Fix async dependency resolution error plus code cleanup
Async dependency resolution: sometimes UDP queue and binding could started in a different order bc of the Guava ServiceManager that starts everything asynchronuously. If the queue was started after the binding, no multicast was received. Fixed by explicitly setting the queue as a message receiver for the binding before anything is started.