Refactor ServiceId from URI to String
Rationale: according to dpws:4.2 Hosting ServiceId is compared using simple string comparison: Identifier for a HOSTED SERVICE which MUST be persisted across re-initialization and MUST NOT be shared across multiple Hosted elements. ServiceId MUST be unique within a DEVICE. This value should be compared directly, as a case-sensitive string, with no attempt to make a relative URI into an absolute URI, to unescape, or to otherwise canonicalize it.