Fix JAXB concurrency problem
Create marshaller and unmarshaller on every request. Rationale: according to https://stackoverflow.com/questions/7400422/jaxb-creating-context-and-marshallers-cost:

"JAXBContext is thread safe and should only be created once and reused to avoid the cost of initializing the metadata multiple times.  Marshaller and Unmarshaller are not thread safe, but are lightweight to create and could be created per operation."